requirements.txt
openai
elasticsearch>=8.0.0
PyYAML
Jinja2
python-dateutil
